Moved git specific CI preparation tasks into separate plugin
1. This way, it is cleaner.
The GitPlugin was becoming a bottomless bucket for all Git tasks.
This should be also easier to setup with other CI servers,
e.g. when we need to introduce a JenkinsPlugin or something like that.
2. Public API change: travisReleasePrepare -> ciReleasePrepare
